+++ src/app/app-routing.module.ts
@@ -71,6 +71,10 @@ const routes: Routes = [
     loadChildren: () => import('./features/events/events.module').then(m => m.EventsModule)
   },
   {
+    path: 'audit',
+    loadChildren: () => import('./features/audit/audit.module').then(m => m.AuditModule)
+  },
+  {
     path        : '**',
     pathMatch   : 'full',
     loadChildren: () => import('./pages/error404/error404.module').then(m => m.Error404Module)