+++ src/app/features/audit/audit.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+
+import { SharedModule } from '@shared/shared.module';
+
+import { AuditRoutingModule } from '@features/audit/audit-routing.module';
+import { AuditComponent } from '@features/audit/audit.component';
+
+@NgModule({
+  declarations: [
+    AuditComponent
+  ],
+  imports: [
+    SharedModule,
+    AuditRoutingModule
+  ]
+})
+export class AuditModule { }