+++ src/app/features/users/users.module.ts
@@ -7,7 +7,7 @@ import { UsersComponent } from '@features/users/users.component';
 import { UserViewComponent } from '@features/users/user-view/user-view.component';
 import { UserEditComponent } from '@features/users/user-edit/user-edit.component';
 import { UserAdminEntitiesComponent } from '@features/users/user-admin-entities/user-admin-entities.component';
-import { UserAuditComponent } from './user-audit/user-audit.component';
+import { UserAuditComponent } from '@features/users/user-audit/user-audit.component';
 
 @NgModule({
   declarations: [